# docs/api/management-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GEOFlow remote management preview",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/tasks/{task}/enqueue": {
            "post": {
                "operationId": "tasks.enqueue",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "x-required-scope": "tasks:write",
                "x-idempotent": true,
                "x-receipt": true,
                "responses": {
                    "200": {
                        "description": "Successful API envelope"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Baseline or request identity conflict"
                    },
                    "201": {
                        "description": "Resource or receipt created"
                    }
                },
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}